moved the versionProvider into a seperate class for easy access
