+++ src/main/java/de/fraunhofer/aisec/codyze/ManifestVersionProvider.java
@@ -1,0 +1,40 @@
+package de.fraunhofer.aisec.codyze;
+
+import picocli.CommandLine;
+
+import java.io.IOException;
+import java.net.URL;
+import java.util.Enumeration;
+import java.util.jar.Attributes;
+import java.util.jar.Manifest;
+
+public class ManifestVersionProvider implements CommandLine.IVersionProvider {
+    // adapted example from https://github.com/remkop/picocli/blob/master/picocli-examples/src/main/java/picocli/examples/VersionProviderDemo2.java
+    @Override
+    public String[] getVersion() throws Exception {
+        Enumeration<URL> resources = CommandLine.class.getClassLoader().getResources("META-INF/MANIFEST.MF");
+        while (resources.hasMoreElements()) {
+            URL url = resources.nextElement();
+            try {
+                Manifest manifest = new Manifest(url.openStream());
+                if (isApplicableManifest(manifest)) {
+                    Attributes attr = manifest.getMainAttributes();
+                    return new String[] { get(attr, "Implementation-Version").toString() };
+                }
+            }
+            catch (IOException ex) {
+                return new String[] { "Unable to read from " + url + ": " + ex };
+            }
+        }
+        return new String[] { "Unable to find manifest file." };
+    }
+
+    private boolean isApplicableManifest(Manifest manifest) {
+        Attributes attributes = manifest.getMainAttributes();
+        return "codyze".equals(get(attributes, "Implementation-Title"));
+    }
+
+    private static Object get(Attributes attributes, String key) {
+        return attributes.get(new Attributes.Name(key));
+    }
+}
